Make Point() return Integer again (reverting/replacing 12ff8fec)
This patch changes fb_GfxPoint() to return an Integer (again), instead of
a Long that would sometimes be wrapped in a Culng(). Furthermore it reverts
the rgb() and rgba() macros back to returning UIntegers, instead of ULongs.

Doing the implicit Culng() around Point() when retrieving colors was bad,
because it broke backwards compatibility by changing the outcome of
calculations involving Point().

So here's a better way to fix the issue where an RGBA color returned
through a signed 32bit value could be sign-extended on 64bit and then
compare unequal to the non-sign-extended version of that RGBA color:

Simply return it as a signed 64bit value on 64bit! The gfxlib2 side has to
take care not to do sign-extension accidentally in fb_GfxPoint() but that's
ok. Plus, on the 32bit port, this doesn't even change the ABI:
Integer/Long (a.k.a. ssize_t/int) are the same thing there.

Using Integer may not make the most sense for RGBA colors, but Point() can
also return palette indices in other screen modes, and Integer makes sense
for those.