Determine struct return type after adding implicit members
This fixes an issue where a structure could be seen as trivial despite
non-trivial: hGetReturnType() calls symbCompIsTrivial(), but that depends
on the implicit members being added already or it could produce wrong
results.

The included test case exposes the problem by triggering an assert(),
but only on Win32, because it was trying to return the struct in registers
despite that not being possible for this struct. And it didn't happen on
other systems because those never return structs in registers at all.