parser: Don't abort when warning about mismatching parameter initializers
Since the "mismatching parameter initializers" warning was added in
301962fc, it also caused hCheckPrototype() to abort, which caused
cProcHeader() to do "return CREATEFAKE( )" which it normally only does
for error recovery.

1. This should just be warning though, not an error.
2. The symbol created by CREATEFAKE() is not completely the same as what
would be produced in the absence of errors; for example it will use a
temporary identifier instead of the original. Because of this, the newly
added "warning" could even cause procedure bodies to be emitted under an
incorrect name, resulting in linker errors.