overload resolution: Ingore CONSTness for RTL procedures without RTL_CONST
(closes #735)

This is a regression from 8d8df7ad. Before that commit, the code contained
two calls to symbCheckConstAssign(), the first guarded by a
	if( typeGetConstMask( param_dtype ) ) then
and the second guarded by
	if( (symbGetIsRTL( parent ) = FALSE) or (symbGetIsRTLConst( param )) ) then
Given that none of the RTL functions had CONSTs on their parameters,
all were redirected to the second check, causing them to be handled
correctly. The commit changed the code though, such that the first call
to symbCheckConstAssign() was no longer guarded this way.