overload resolution: Combine and fix redundant CONSTness checks
(closes #728)

1. The first symbCheckConstAssign() did not consider parameter mode

2. The entire code path/logic seemed weird/wrong - e.g. two calls to
symbCheckConstAssign(), one outside the "type_is_compatible" guard.

This already doesn't make sense as it would allow incompatible types to
match during overload resolution if they match based on CONSTness alone.

Also, the part of code using const_matches for the second
symbCheckConstAssign() call does not make sense, as it can never be reached
if the types are compatible, because then the first constness check will
already have determined the match. Thus for now that is removed too, though
I think it can be fixed in the future.