Implement implicit UDT members after calculating UDT return type
This fixes an issue where the implementation of implicit UDT members
(specifically, building the vtable) wanted to use the UDT return type
(due to a virtual method with byval self function result), but could not
because it was not known yet (it triggered the assert() in
symbProcRecalcRealType() when creating the function pointer for the
vtable slot).

This is a regression from 8b5db5c1 where the order of
adding the implicits and calculating the UDT return type was changed.

The solution to fix both cases (the original issue from the old commit,
and the regression introduced by the old fix) is to split declaration from
implementation of implicit members, and calculate the UDT return type in
between.