ASM backend: Adjust emitVARINIi()'s value parameter to be a LONGINT
This way the str() it does will be a str(longint) instead of a
str(integer), i.e. it will produce the same output no matter whether
on 32bit or 64bit. E.g. &hFFFFFFFF in variable initializers will now
always be emitted as 4294967295 instead of -1 (with 32bit signed Integer)
or 4294967295 (with 64bit signed Integer).

This matches the other code dealing with raw values (e.g. FBVALUE or
IR's _emitVarIniI()).