Combine symbAllocIntConst() and symbAllocLongIntConst() in similar way
The two were exactly the same except for the value parameter type which is
only used as input to the hex() call, resulting in hex(longint) in
symbAllocLongIntConst() (intended for 64bit values) and hex(integer) in
symbAllocIntConst() (intended for 32bit values).

Not only is that duplicate code, using Integer for 32bit is also suboptimal
since it changes to 64bit on 64bit. So now the combined function uses
hex(longint) and hex(long) for 64bit/32bit values respectively.