Remove contrib/bootstrap.mk
It was intended to partially automate the process of "precompiling fbc
sources to *.asm or *.c files, and then compiling that into a new fbc
after they were taken to the target system".

This is now documented at
	http://www.freebasic.net/wiki/wikka.php?wakka=DevBootstrap

contrib/bootstrap.mk can be removed because it's outdated (e.g. still using
bin/fbc-new which FB's makefile no longer does) and only worked with 32bit
x86 Linux. It would have to be adjusted to support 64bit and other systems.

Overall though I don't think it's worth it. This script overcomplicates
things in comparison with the couple simple commands needed to actually
bootstrap fbc... so the script mainly is an unnecessary maintenance burden.