ASM backend: Don't allocate stack space for fake dynamic array symbols
Fixed: #739

Only the array descriptors should be allocated on stack, not the internal
dynamic array symbols. The C/LLVM backends already do this correctly.

The ASM backend though reserved extra memory based on the dynamic array
symbol's FBSYMBOL.lgt field, which is set to the size of the array element
data type by symbAddVar(). This memory was never used though, since the
array descriptor was still allocated separately, and all accesses to the
array go through the descriptor anyways. In the worst case it could cause
a stack overflow.