Remove contrib/libs/ build scripts
As mentioned in
	http://www.freebasic.net/forum/viewtopic.php?f=17&t=22802
I think that the fbc project cannot provide precompiled external libraries
out-of-the-box. It's not a Linux distro and doesn't have the manpower.
Thus I no longer want to do it, in favour of much more easier/frequent
FB releases.

Nevertheless, if someone wanted to revive the idea, to help FB users,
they can still do so - the build scripts are still in the Git history...
It should be noted though that these scripts are outdated (from 2013),
and any such future project may want to consider a different approach.