Set FB version to 1.00.0
As discussed on the forum, we're going to release FB 1.0 soon. I think the
compiler is in a good shape right now. It has tons of fixes and some nice
new features including the 64bit support, thread-safe gfxlib, and dynamic
arrays in UDTs ;)

It's a good time to declare it 1.0 (as was already planned for 2014 since
the 0.90 release in 2013), because of the 64bit support which essentially
concludes the -gen gcc work done since FB 0.21 times and brought some huge
changes to the FB world.

Besides, FB is used as if it is 1.0 since a long time. Sure there are a
lot of known issues (mostly rare edge cases) remaining, but that's nothing
special, as there always are bugs. Since FB 0.24, a lot of missing
features were added, hence the version jumps to 0.90 and 1.00.