Unix rtlib: Make __fb_bg_mutex recursive like the others
This should help prevent dead-locks in case a signal handler from
src/rtlib/signals.c will be called (e.g. in case of SIGINT), because it
calls fb_End() -> fb_hEnd() -> fb_hExitConsole() which may attempt to
acquire the __fb_bg_mutex recursively. Of course this is not a full fix
for the broken signal handler (it should not invoke any non-async-signal-
safe functions).