parser: Make WITH + compound stack work like most other compound blocks
Previously, parser.stmt.with held the current WITH context, while the nodes
on stack represented the previous ones.

Now it works like most other compound blocks, i.e. parser.stmt.with is just
a pointer into the stack.

This gives us access to other fields of the stack nodes, such as
FB_CMPSTMTSTK.scopenode, which is not part of FB_CMPSTMT_WITH. In other
words, now it becomes possible to give WITH a temporary scope using the
same code as for many other compounds.