Fix Hiword() for 64bit
Now there are two versions of Hiword(), one for 32bit and one for 64bit.

The 64bit version needs an extra operation to mask out the upper dword of
the 64bit Uinteger operand.

The 32bit version doesn't need that though, and I think it's best to keep
it separate from the 64bit version, in order to avoid the extra operation
which is unnecessary on 32bit and may not be optimized out.