overload: Allow passing constant 0 Byte/Short to pointer params
This makes the behaviour of normal and overloaded calls match.
ast-node-conv allows assigning a literal zero integer to a pointer even if
the integer type's size didn't match the pointer size. hCalcTypesDiff() on
the other hand had a check only allowing this if the type size matched.

To fix this, the non-ptr-to-ptr checking code from ast-node-conv.bas is
extracted into a helper function which is now called by hCalcTypesDiff(),
replacing the bogus size check. This should ensure that both overload
resolution and normal cast/conv checks will behave the same (as far as
passing literal zeroes to pointers is concerned).

This also fixes the issue where it was no longer possible to call Bsave
in -lang qb with 0 as the 2nd parameter (which is a pointer), since commit
07d6977397 in FB 0.90 which turned Bsave() into an overloaded function.