ast: Fold @x+N pointer-arithmetic BOPs into VARs with offsets
This adds a folding optimization to astNewBOP(). Given a BOP such as
	BOP(+, ADDROF(VAR(x)), CONST(N))
we can encode the N offset into the VAR node and remove the BOP.
	ADDROF(VAR(x, ofs=N))
which allows better code generation later.

@x + N expressions are rather common, not necessarily in FB code, but
definitely for array and field accesses as produced behind the scenes in
the compiler.

However, this optimization wasn't really "missing" in FB versions before
1.00.0, because the compiler directly built VARs with offsets for array
accesses. Thus this optimization wasn't necessary.

But since the changes in 1.00.0 (dynamic array fields etc.) the compiler
now builts the pointer-arithmetic-style BOPs instead. Now this
optimization was missing, and thus the generated code for affected cases
was worse. (I noticed this with the field accesses generated to fill array
descriptors)

Instead of adjusting the compiler to produce VARs with offsets for field
accesses etc. as it did before, I think it's better to implement this
generic folding which will also improve code in other cases.

As a side-effect, this changes the "typical" internal representation of
field accesses. For those, we build the following:
	FIELD(DEREF(BOP(+, ADDROF(VAR(udt)), offsetof(field))))
which due to the change in astNewBOP() will become:
	FIELD(DEREF(ADDROF(VAR(udt, offsetof(field)))))
and then astNewDEREF will also cancel out the DEREF/ADDROF:
	FIELD(VAR(udt, offsetof(field)))
i.e. field accesses are now represented in a much more natural way.
This affects code checking for the old field access pattern; I found two:
1. astIsAccessToLocal(): This is easy to adjust to the new pattern.
2. hMergeNestedFIELDs(): Nested field accesses (udt.a.b) now look like:
	FIELD(FIELD(VAR(udt, offsetof(a) + offsetof(b))))
   i.e. the hMergeNestedFIELDs() optimization is completely useless now,
   as the field offsets are accumulated thanks to the new optimization in
   astNewBOP() already. The only purpose that hMergeNestedFIELDs() still
   serves is to remove the nested FIELD nodes (though even that isn't
   really needed at all, or at the very least could be moved into
   astNewFIELD()).