parser: Allow ignoring results of non-trivially type-casted function calls
Previously it was only allowed to ignore the results of function calls
wrapped in noconv casts, e.g.
	cuint(returnAnInteger())

Now it becomes possible to ignore real conversions, e.g.
	cdbl(returnAnInteger())
	clng(returnALongInt())
	clngint(returnAByte())
	etc.

This will be especially useful for the new 64bit-capable Windows API
headers I'm working on, because they contain a lot of such cases:
	cast(LONG, SendMessage(...))
where LONG is a 32bit Long and SendMessage returns an Integer-sized type,
specifically Long on 32bit and LongInt on 64bit. On 32bit such casts are
noconv casts and were already allowed, but on 64bit they're non-trivial
casts (converting a 64bit integer to a 32bit one), so weren't allowed yet.

As is usual with the Windows API, a lot of such functions return an error
code, which users may not always want to check. Thus I think it should be
possible to ignore these function results even if they're type-casted, as
in C.

Adjusting the headers to allow it is much more work than adjusting the
compiler, and sometimes potentially impossible (in cases where the function
result needs to be type-casted if you want to check it).