tests/warnings: Test for multiple targets at once
This changes the warning tests over to testing for multiple targets (dos,
linux-x86, linux-x86_64, win32, win64) at once, instead of just for the
host.

This is necessary because fbc can produce different warnings in each case.
(differences between systems, or 32bit vs 64bit)

Also it's good to be able to test all targets, regardless of the host
system (ensuring cross-compiling produces the same results as native
compilation for the target).

To implement this, I added a shell script instead of continuing to "abuse"
a makefile as one. We always run all tests here, there is no
dependency-based rebuilding.