Don't treat base/derived UDT as equal for BYVAL results/parameters
Function pointer assignments (or virtual overriding) where one function
returns the base UDT BYVAL and the other returns a derived UDT BYVAL, we
can't treat them as equal. For example:

	type A extends object
		declare virtual function f( ) as A
	end type
	type B extends A
		declare function f( ) as B override
	end type

A/B are non-trivial UDTs, so they will be returned in a temp var on stack.
I.e. when calling A.f, the caller will allocate room for an A, but if we
end up calling B.f (due to virtual + overriding), it would try to write a
B object into the A temp var - potential buffer overflow, etc.

In other words, in general, the calling conventions don't allow covariance
for BYVAL function results. It's only allowed with BYREF results. This is
also how it works with g++/clang++.

The same more or less also applies to parameters:

	type A
		i as integer
	end type
	type B extends A
		j as integer
	end type

	sub f( byval xb as A )
	end sub

	dim p as sub( byval as B ) = @f

It seems like this is not safe in general either. Here, f() expects an A
object on the stack (BYVAL, plain-old-data), but calling p() will push a
B object to the stack. At least with stdcall, f wouldn't clean up the
stack properly (because it only expected an A, it will only pop sizeof(A)
bytes from the stack, instead of sizeof(B)).
g++/clang++ also disallow this.