inc: New Windows API headers based on MinGW-w64 3.3.0
The new headers are based on MinGW-w64 instead of MinGW.org headers,
because MinGW.org doesn't support 64bit yet. This alone causes lots of
differences due to the differences between MinGW-w64 and MinGW.org.

Some headers don't exist under MinGW-w64; some declarations are moved
around; there are lots of new files. The base API (winbase/winuser/
commctrl/etc.) is pretty much the same, but differences increase in other
places such as DirectX headers.

One fundamental difference in these new headers is that the *A/*W
declaration pattern changed. Previously we used:
	#ifdef UNICODE
		declare foo alias "fooW"
	#else
		declare foo alias "fooA"
	#endif

now it's as in the original Windows SDK:
	declare sub fooW
	declare sub fooA
	#ifdef UNICODE
		#define foo fooW
	#else
		#define foo fooA
	#endif

The main reason for this is that the headers are auto-generated with fbfrog
and it'd be too complicated to change the code around to match the old
pattern. Of course the new pattern allows you to use *A/*W functions
directly/explicitly, which can be seen as advantage too.

The new headers support 32bit and 64bit, UNICODE and non-UNICODE, and three
hand-selected _WIN32_WINNT settings: &h0400 &h0502 &h0602. More could be
done in the future but fbfrog is too slow for it at the moment. With these
three settings though we at least support the main three Windows API
levels: older Windows NT, Windows XP, and new Windows Vista/7/8.

DDK: The headers in inc/win/ddk/ aren't updated yet, but could be done in
the future.

GdiPlus: The C++ API isn't updated because fbfrog doesn't support C++ yet,
but the flat C API is updated. Hence I left the existing Gdi[P|p]lus*.bi
files in place, added the new gdiplus-c.bi, which will be used by
GdiPlus.bi for 64bit. This means we don't have GdiPlus C++ headers for
64bit yet, but at least the C part.