SSE emitter: Fix potential ICE if certain float constants exist
This fixes an issue where the SSE emitter would try to add global integer
vars with certain values, but fail because a global var for that value
already existed.

Example code that triggered the compiler crash under -fpu sse -fpmode fast:
	const pi_2 = 1.570796326794
	dim as single a, b
	a = PI_2
	b = cos(a)

Reported at: http://www.freebasic.net/forum/viewtopic.php?f=3&t=23280

The problem is that symbAllocFloatConst/symbAllocIntConst can produce
global vars with the same internal name, but fail to look them up if the
dtype is different. Thus they would try to re-add a global var with the
same name, instead of re-using the existing one. Then the symbAddVar()
will fail because of the "duplicate definition", and the symbAlloc*Const
functions don't expect/handle that case.

symbAllocFloatConst uses FB_SYMBOPT_NODUPCHECK to work-around that issue,
and the simple fix for the compiler crash issue is to make symbAllocIntConst
use that too.

But symbAllocIntConst also failed to properly differentiate between vars
with same value but different type size (because it only encoded the hex
value without leading zeroes, thus not the type size). That needs to be
fixed too.

At that point another change can be made: the lookup to reuse existing
vars can be done via symbLookupByNameAndClass instead of
symbLookupByNameAndSuffix. This means that dtype no longer matters.
Assuming that value and type size are properly encoded in the var's name,
that's all we need to get the proper var. It doesn't matter whether we
have a Single or Long (or Double/Longint), as long as the byte values are
the same.

Further changes:
* common code from symbAllocFloatConst/symbAllocIntConst extracted into
  hAllocIntOrFloatConst
* symbLookupByNameAndSuffix removed, because it's no longer used

This issue showed up since commit 42e72dcf5e, which changed hFloatToStr()
and thus affected the internal names used by symbAllocFloatConst.