non-standalone: Use 8.3 dir layout if on DOS, not if compiling for DOS
This reverts commit 8785bd816ccb6598e4219ce10cc72820a4a74316.

The decision whether to use the 8.3-filename-format-compatible dirlayout
(include/freebas/ and lib/freebas/ instead of include/freebasic/ and
lib/freebasic/) has to be made based on the host system, not based on the
compilation target.

1. The 8.3 names are needed on DOS only, but for all targets (although of
course in practice probably nobody is cross-compiling from DOS to Linux)

2. When cross-compiling from other systems to DOS, the 8.3 names shouldn't
be needed, because having to have both dir layouts is rather annoying. It's
not that bad to have lib/freebas/dos next to lib/freebasic/linux, but it's
annoying that headers are searched in include/freebas/ instead of
include/freebasic/.

The behaviour of the FB makefile is currently still broken in this regard,
because it puts the DOS libs into the lib/freebas/dos/ subdir no matter
what the host system is. Changing this probably requires introducing
HOST_OS/HOST_ARCH variables separate from TARGET_OS/TARGET_ARCH, and this
seems pretty complex to me, so it's probably better to leave that to the
future.