inc: Turn many #defines into constants
In general constants are better because they're less intrusive - e.g.
#defines can conflict with fields or parameters, but constants can't.

For example, with OPAQUE from win/wingdi.bi being a constant instead of a
#define, there no longer is a conflict with zlib.bi's "opaque" fields/
parameters.

And since FB's #if supports constants too nowadays, that's one less reason
to stick with #defines.

The only case where a #define is needed is in cases where the macro
expansion is needed, e.g. if such #defines are supposed to be used with
the # or ## preprocessing operators.

* This changes #defines that represent simple expressions such as integer
  literals over to constants
* constants referencing other constants are not recognized as constants yet
  by fbfrog (and I wouldn't want to do it manually), but this could be
  improved in the future.